after attack
https://github.com/tgstation/tgstation/pull/72320 https://github.com/tgstation/tgstation/pull/44409

+++ code/game/objects/items/crayons.dm
@@ -259,13 +259,11 @@
 	var/static/regex/crayon_r = new /regex(@"[^\w!?,.=%#&+\/\-]")
 	return replacetext(LOWER_TEXT(text), crayon_r, "")
 
-/obj/item/toy/crayon/afterattack(atom/target, mob/user, proximity, params)
-	. = ..()
-	if(!proximity || !check_allowed_items(target))
-		return
+/// Attempts to color the target. Returns how many charges were used.
+/obj/item/toy/crayon/proc/use_on(atom/target, mob/user, params)
 
 	if(is_capped)
-		return
+		return 0
 
 	var/static/list/punctuation = list("!","?",".",",","/","+","-","=","%","#","&")
 
@@ -283,7 +281,7 @@
 		target = target.loc
 
 	if(!isValidSurface(target))
-		return
+		return 0
 
 	var/drawing = drawtype
 	switch(drawtype)
@@ -395,7 +393,7 @@
 	// hippie start -- using changes moved to the end of the proc, so it won't use charges if the spraying fails for any reason.
 	var/charges_used = use_charges(user, cost)
 	if(!charges_used)
-		return
+		return 0
 	. = charges_used
 	// hippie end
 	var/fraction = min(1, . / reagents.maximum_volume)
@@ -405,6 +403,22 @@
 		reagents.expose(t, TOUCH, fraction * volume_multiplier)
 		reagents.trans_to(t, ., volume_multiplier, transfered_by = user)
 	check_empty(user)
+	return .
+
+/obj/item/toy/crayon/afterattack(atom/target, mob/user, proximity, params)
+	. = ..()
+
+	if(!proximity)
+		return .
+
+	if (isitem(target))
+		. |= AFTERATTACK_PROCESSED_ITEM
+
+	if (!check_allowed_items(target))
+		return .
+
+	use_on(target, user, params)
+	return .
 
 /obj/item/toy/crayon/attack(mob/M, mob/user)
 	if(edible && (M == user))
@@ -648,18 +662,15 @@
 		. += "It is empty."
 	. += span_notice("Alt-click [src] to [ is_capped ? "take the cap off" : "put the cap on"].")
 
-/obj/item/toy/crayon/spraycan/pre_attack(atom/target, mob/user, proximity, params)
-	if(!proximity)
-		return ..()
-
+/obj/item/toy/crayon/spraycan/use_on(atom/target, mob/user, params)
 	if(is_capped)
 		if(is_type_in_typecache(target, spraycan_touch_normally) || target.GetComponent(/datum/component/storage))
 			return ..()
-		to_chat(user, span_warning("Take the cap off first!"))
-		return
+		balloon_alert(user, "take the cap off first!")
+		return 0
 
 	if(check_empty(user))
-		return
+		return 0
 
 	if(iscarbon(target))
 		if(pre_noise || post_noise)
@@ -685,7 +696,7 @@
 		var/fraction = min(1, . / reagents.maximum_volume)
 		reagents.expose(C, VAPOR, fraction * volume_multiplier)
 
-		return
+		return .
 
 	if(isobj(target) && !(target.flags_1 & UNPAINTABLE_1))
 		if(actually_paints)
@@ -726,7 +737,7 @@
 		user.visible_message("[user] coats [target] with spray paint!", span_notice("You coat [target] with spray paint."))
 		return FALSE
 
-	. = ..()
+	return ..()
 
 /obj/item/toy/crayon/spraycan/update_icon_state()
 	icon_state = is_capped ? icon_capped : icon_uncapped
@@ -745,11 +756,20 @@
 	charges = -1
 
 /obj/item/toy/crayon/spraycan/borg/afterattack(atom/target,mob/user,proximity, params)
-	var/diff = ..()
+	if (!proximity)
+		return
+
+	if (isitem(target))
+		. = AFTERATTACK_PROCESSED_ITEM
+
+	if (!check_allowed_items(target))
+		return .
+
+	var/diff = use_on(target, user, params)
 	if(!iscyborg(user))
 		to_chat(user, span_notice("How did you get this?"))
 		qdel(src)
-		return FALSE
+		return .
 
 	var/mob/living/silicon/robot/borgy = user
 
@@ -762,6 +782,8 @@
 	// it's free.
 	if(borgy.cell)
 		borgy.cell.use(cost)
+
+	return .
 
 /obj/item/toy/crayon/spraycan/hellcan
 	name = "hellcan"
